Move helper functions used by update_head_node recipe into it
Add Rspec setup to slurm cookbook to permit to mock helpers functions
